src/main/java/com/speakplusplus/onlinequizwebservice/config/EmailConfirmationProps.java
package com.speakplusplus.onlinequizwebservice.config;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.context.annotation.Configuration;

@Configuration
@ConfigurationProperties(prefix = "email.confirmation")
public class EmailConfirmationProps {

    private String host;
    private String endpoint;

    public String getHost() {
        return host;
    }

    public void setHost(String host) {
        this.host = host;
    }

    public String getEndpoint() {
        return endpoint;
    }

    public void setEndpoint(String endpoint) {
        this.endpoint = endpoint;
    }
}
